third_party/blink/web_tests/virtual/css-anchor-update-disabled/external/wpt/css/css-anchor-position/position-area-in-position-try-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Placement: --right flip-x
  assert_equals: offsetLeft expected 40 but got 350
[FAIL] Placement: --bottom flip-y
  assert_equals: offsetLeft expected 95 but got 350
[FAIL] Placement: --left flip-x, --top
  assert_equals: offsetLeft expected 95 but got 350
[FAIL] Placement: --top flip-y, --left
  assert_equals: offsetLeft expected 40 but got 350
[FAIL] Placement: --left flip-start flip-y, --left
  assert_equals: offsetLeft expected 40 but got 350
Harness: the test ran to completion.
